backend/requirements.txt
Django==4.0
djangorestframework==3.13.1
psycopg2-binary==2.9.3
djangorestframework-simplejwt==5.2.0
firebase-admin==5.2.0
stripe==4.0.0
django-cors-headers==3.10.0
python-dotenv==0.19.2